Escape Go templates (#1943)
